requirements.txt
Flask==2.2.2
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
mysqlclient
sqlalchemy-utils
WTForms
email-validator
python-dotenv
